requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
Django==3.1.5
django-autoslug==1.9.8
django-taggit==1.3.0
django-tagulous==1.1.0
djangorestframework==3.12.2
Faker==5.0.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
npm==0.1.1
optional-django==0.1.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
